requirements.txt
Flask==2.3.3
pymysql==1.1.0
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
redis==5.0.1  # Cho detection system
pyarrow==14.0.1  # Cho Parquet support
requests==2.31.0  # Cho attack simulator
# Data Generation Requirements
PyYAML>=6.0
tqdm>=4.66.0  # For progress bars